refactor: extract memo dialog
